move out components to see if it helps server func count
