+++ src/(components)/about.mdx
@@ -1,4 +1,4 @@
-import { Card } from "../(components)/common/card";
+import { Card } from "./common/card";
 
 <Card>
 <article className="prose p-8">
