+++ src/app/(core)/page.tsx
@@ -1,4 +1,4 @@
-import About from "../(components)/about.mdx";
+import About from "~/(components)/about.mdx";
 
 export default async function Home() {
   return (
