+++ src/app/error.tsx
@@ -1,5 +1,5 @@
 "use client";
 
-import { ErrorPage } from "./(components)/error-page";
+import { ErrorPage } from "~/(components)/error-page";
 
 export default ErrorPage;
